secure-file-share/backend/requirements.txt
# Core
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Security
pyotp==2.9.0
cryptography==42.0.2

# Development
python-dotenv==1.0.1
pytest==8.0.0
pytest-django==4.8.0
black==24.1.1
flake8==7.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0 